Fix 15 or 31 else high count player regression
In play versus play highest count this fix restores 0.09 of the 0.15
points per hand advantage for pone and 0.17 of 0.56 of the points per
hand advantage to dealer.

This fix also introduces type hints in areas where helpful and mypy to
the project dependencies.  (Note: not all pre-existing mypy-flagged
errors fixed - yet.)