Reify entire play in TypeScript simulations
Has an approximately 20% performance cost but without it simulation
code diverges from UI code which requires this state information to be
explicit.  (And divergence means double the code to maintain, which is
not so affordable in a single part-time developer project!)  Code also
likely will make adding of test automation much easier and will be
required to enable easy to test and implement simulations starting
mid-play and undo.